代码优化

+++ app/src/main/java/com/pengxh/autodingding/ui/HistoryActivity.java
@@ -1,12 +1,14 @@
 package com.pengxh.autodingding.ui;
 
+import android.os.Environment;
+import android.util.Log;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.ListView;
 import android.widget.RelativeLayout;
 
 import com.aihook.alertview.library.AlertView;
-import com.aihook.alertview.library.OnItemClickListener;
+import com.google.gson.Gson;
 import com.gyf.immersionbar.ImmersionBar;
 import com.pengxh.app.multilib.base.BaseNormalActivity;
 import com.pengxh.app.multilib.utils.DensityUtil;
@@ -14,19 +16,20 @@ import com.pengxh.app.multilib.widget.EasyToast;
 import com.pengxh.autodingding.R;
 import com.pengxh.autodingding.adapter.HistoryAdapter;
 import com.pengxh.autodingding.bean.HistoryBean;
+import com.pengxh.autodingding.utils.ExcelUtils;
 import com.pengxh.autodingding.utils.SQLiteUtil;
 import com.pengxh.autodingding.utils.Utils;
 import com.pengxh.autodingding.widgets.EasyPopupWindow;
 import com.scwang.smartrefresh.layout.SmartRefreshLayout;
 
+import java.io.File;
 import java.util.Arrays;
 import java.util.List;
 
 import butterknife.BindView;
 import butterknife.OnClick;
 
-public class HistoryActivity extends BaseNormalActivity
-        implements View.OnClickListener, OnItemClickListener {
+public class HistoryActivity extends BaseNormalActivity implements View.OnClickListener {
 
     @BindView(R.id.titleLayout)
     RelativeLayout titleLayout;
@@ -37,8 +40,8 @@ public class HistoryActivity extends BaseNormalActivity
     @BindView(R.id.historyList)
     ListView historyList;
 
+    private static final String TAG = "HistoryActivity";
     private static final List<String> items = Arrays.asList("删除记录", "导出记录");
-    private AlertView alertView;
     private List<HistoryBean> historyBeans;
     private HistoryAdapter historyAdapter;
     private SQLiteUtil sqLiteUtil;
@@ -79,10 +82,14 @@ public class HistoryActivity extends BaseNormalActivity
                     new AlertView("温馨提示", "空空如也，无法删除", null, new String[]{"确定"}, null, this, AlertView.Style.Alert,
                             null).setCancelable(false).show();
                 } else {
-                    sqLiteUtil.deleteAll();
-                    historyBeans.clear();
-                    historyAdapter.notifyDataSetChanged();
-                    emptyView.setVisibility(View.VISIBLE);
+                    new AlertView("温馨提示", "是否确定清除打卡记录？", "取消", new String[]{"确定"}, null, this, AlertView.Style.Alert, (o, position1) -> {
+                        if (position1 == 0) {
+                            sqLiteUtil.deleteAll();
+                            historyBeans.clear();
+                            historyAdapter.notifyDataSetChanged();
+                            emptyView.setVisibility(View.VISIBLE);
+                        }
+                    }).setCancelable(false).show();
                 }
             } else if (position == 1) {
                 String emailAddress = Utils.readEmailAddress();
@@ -94,26 +101,28 @@ public class HistoryActivity extends BaseNormalActivity
                     EasyToast.showToast("无打卡记录，无法导出", EasyToast.WARING);
                     return;
                 }
-                alertView = new AlertView("温馨提示", "导出到" + emailAddress, "取消", new String[]{"确定"}, null, this, AlertView.Style.Alert,
-                        this).setCancelable(false);
-                alertView.show();
+                new AlertView("温馨提示", "导出到" + emailAddress, "取消", new String[]{"确定"}, null, this, AlertView.Style.Alert, (o, position1) -> {
+                    if (position1 == 0) {
+                        //导出Excel
+                        pullToEmail(historyBeans);
+                    }
+                }).setCancelable(false).show();
             }
         });
         easyPopupWindow.showAsDropDown(titleLayout, titleLayout.getWidth(), DensityUtil.dp2px(this, 1));
     }
 
-    @Override
-    public void onItemClick(Object o, int position) {
-        switch (position) {
-            case -1:
-                alertView.dismiss();
-                break;
-            case 0:
-                //导出Excel
-                EasyToast.showToast("预留功能，目前暂无", EasyToast.WARING);
-                break;
-            default:
-                break;
+    private static final String[] title = {"uuid", "日期", "打卡信息"};
+
+    private void pullToEmail(List<HistoryBean> historyBeans) {
+        //{"date":"2020-04-15","message":"考勤打卡:11:42 下班打卡 早退","uuid":"26btND0uLqU"},{"date":"2020-04-15","message":"考勤打卡:16:32 下班打卡 早退","uuid":"UTWQJzCfTl9"}
+        Log.d(TAG, "pullToEmail: " + new Gson().toJson(historyBeans));
+        File dir = new File(Environment.getExternalStorageDirectory(), "DingRecord");
+        if (!dir.exists()) {
+            dir.mkdir();
         }
+        ExcelUtils.initExcel(dir.toString() + "/打卡记录表.xls", title);
+        String fileName = Environment.getExternalStorageDirectory() + "/DingRecord/打卡记录表.xls";
+        ExcelUtils.writeObjListToExcel(historyBeans, fileName, this);
     }
 }
